requirements.txt
paho-mqtt==1.2
pyserial==3.2.1
